refactor lib.rs to compile.rs & calls.rs
