+++ packages/vm/Cargo.toml
@@ -11,8 +11,6 @@ readme = "README.md"
 [dependencies]
 assert_matches = "1.3.0"
 hex = "0.4"
-# parity-wasm = "0.45"
-pwasm-utils = "0.12"
 wasm-instrument = "0.2.0"
 wasmer = { version = "2.3.0", default-features = false, features = ["singlepass", "compiler", "universal"] }
 wasmer-middlewares = "2.3.0"
